requirements.txt
chromadb==0.5.3
langchain_chroma==0.1.3
langchain_community==0.2.16
langchain_core==0.2.38
numpy==2.1.1
pandas==2.0.3
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
sentence_transformers==3.0.1
tenacity==8.5.0
tqdm==4.66.5
